TranslationModel.translate: Always write a translation in the jsonb field
Currently, in the translate function of TranslationModel, the translations
dict is not created if we try to translate the default language.

The suggestion is to always create the translations jsonb field, and
always writing inside, even for the default language.

Thus, all the languages are stored equally and the jsonb field always
contains all the available translations.
If the default language is changed later in the project, the default
translation is kept and still available in the jsonb translations field.

The original attribute is still written in, as a shortcut for accessing
the default language.